Add tests for Resource API

+++ resources/src/controllers/Article.ts
@@ -3,8 +3,13 @@ import mongoose from "mongoose";
 import Article from "../models/Article";
 
 const createArticle = (req: Request, res: Response, next: NextFunction) => {
-  console.log(req.body);
   const { title, body, category } = req.body;
+
+  if (!title || !body || !category) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid data provided" });
+  }
 
   const article = new Article({
     _id: new mongoose.Types.ObjectId(),
@@ -22,10 +27,16 @@ const createArticle = (req: Request, res: Response, next: NextFunction) => {
 const getArticle = (req: Request, res: Response, next: NextFunction) => {
   const articleId = req.params.articleId;
 
+  if (!mongoose.Types.ObjectId.isValid(articleId)) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid ID provided" });
+  }
+
   return Article.findById(articleId)
     .then((article) =>
       article
-        ? res.status(200).json({ article })
+        ? res.status(200).json({ error: false, article })
         : res.status(404).json({ message: "Article not found" })
     )
     .catch((error) => res.status(500).json({ error }));
@@ -33,27 +44,32 @@ const getArticle = (req: Request, res: Response, next: NextFunction) => {
 
 const getAll = (req: Request, res: Response, next: NextFunction) => {
   return Article.find()
-    .then((articles) => res.status(201).json({ articles }))
+    .then((articles) => res.status(200).json({ error: false, articles }))
     .catch((error) => res.status(500).json({ error }));
 };
 
 const getAllTitles = (req: Request, res: Response, next: NextFunction) => {
-  return Article.find({}).select("_id, title, category")
-    .then((articles) => res.status(201).json({ articles }))
+  return Article.find({})
+    .select("_id title category")
+    .then((articles) => res.status(200).json({ error: false, articles }))
     .catch((error) => res.status(500).json({ error }));
 };
 
 const getByCategory = (req: Request, res: Response, next: NextFunction) => {
-  console.log(req.query);
   const categoryParam = req.query?.category;
 
   return Article.find({ category: categoryParam })
-    .then((articles) => res.status(201).json({ articles }))
+    .then((articles) => res.status(200).json({ error: false, articles }))
     .catch((error) => res.status(500).json({ error }));
 };
 
 const updateArticle = (req: Request, res: Response, next: NextFunction) => {
   const articleId = req.params.articleId;
+  if (!mongoose.Types.ObjectId.isValid(articleId)) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid ID provided" });
+  }
 
   return Article.findById(articleId)
     .then((article) => {
@@ -62,10 +78,10 @@ const updateArticle = (req: Request, res: Response, next: NextFunction) => {
 
         return article
           .save()
-          .then((article) => res.status(201).json({ article }))
+          .then((article) => res.status(200).json({ error: false, article }))
           .catch((error) => res.status(500).json({ error }));
       } else {
-        res.status(404).json({ message: "Article not found" });
+        res.status(404).json({ error: true, message: "Article not found" });
       }
     })
     .catch((error) => res.status(500).json({ error }));
@@ -73,11 +89,16 @@ const updateArticle = (req: Request, res: Response, next: NextFunction) => {
 
 const deleteArticle = (req: Request, res: Response, next: NextFunction) => {
   const articleId = req.params.articleId;
+  if (!mongoose.Types.ObjectId.isValid(articleId)) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid ID provided" });
+  }
 
   return Article.findByIdAndDelete(articleId)
     .then((article) =>
       article
-        ? res.status(201).json({ message: "Article Deleted" })
+        ? res.status(200).json({ error: false })
         : res.status(404).json({ message: "Article not found" })
     )
     .catch((error) => res.status(500).json({ error }));
